requirements.txt
# FaceVault — Python dependencies
# Install with:  pip install -r requirements.txt

streamlit>=1.33.0
deepface>=0.0.93
faiss-cpu>=1.8.0        # use faiss-gpu if CUDA is available
numpy>=1.24.0
opencv-python-headless>=4.9.0.80
scikit-learn>=1.3.0     # used internally by DeepFace for clustering metrics
tf-keras>=2.16.0        # Keras backend required by ArcFace model weights

# ── Federated Sync ──
flask>=3.0.0            # lightweight relay server
cryptography>=42.0.0    # Fernet (AES-128-CBC) symmetric encryption
msgpack>=1.0.0          # fast binary serialisation for manifests
matplotlib>=3.8.0       # benchmark plots (ε-accuracy curves, P-R)
requests>=2.31.0        # HTTP client for sync server communication

